Implement AbstractLifecycleEventLoop
